nolong branch finished reviewing all uses of long in OCL.
Now only present in most likely Java-only components.

Interesting issues included:

Canonizer*
 - used to pass bits into the comparison rankings and to sort atom+bond
properties.
 - replaced by int[] with same effect.

Molecule
 - Full list of property flags converted to int.
 - Those with only high-int values renamed *Ex to aid in the
refactoring.
 - The few with both high and low bits renamed *L and *H.
 - high masks (rather than property flags) changed to *H to distinguish
them from individual *Ex propeties.
 - get/setAtomQueryFeature duplicated for *Ex.

long AtomType was not converted to int[], as it is only used in a few
applications.

The performance result is surprisingly high for Javascript.

Running test.OCLSwingJSTest:


JavaScript
DONE 68 1851 ms
Shutdown(0) on testApplet

Java:

DONE 68 13364 ms

Yes, you read that right. JavaScript was 7 times FASTER than Java,
probably because of all the frame openings.

When that frame openings are removed, we get:

JavaScript
DONE 68 1724 ms

and for Java:
DONE 68 1760 ms

+++ src/main/java/com/actelion/research/chem/MolfileCreator.java
@@ -201,7 +201,7 @@ public class MolfileCreator {
 
             mBuilder.append(" 0  0  0");	// massDif, charge, parity
 
-            long hydrogenFlags = Molecule.cAtomQFHydrogen & mol.getAtomQueryFeatures(atom);
+            int hydrogenFlags = Molecule.cAtomQFHydrogen & mol.getAtomQueryFeatures(atom);
             if (hydrogenFlags == 0)
                 mBuilder.append("  0");
             else if (hydrogenFlags == (Molecule.cAtomQFNot0Hydrogen | Molecule.cAtomQFNot1Hydrogen))
@@ -406,7 +406,7 @@ public class MolfileCreator {
             if (no != 0) {
                 int count = 0;
                 for (int atom=0; atom<mol.getAllAtoms(); atom++) {
-                    long ringFeatures = mol.getAtomQueryFeatures(atom) & Molecule.cAtomQFRingState;
+                    int ringFeatures = mol.getAtomQueryFeatures(atom) & Molecule.cAtomQFRingState;
                     if (ringFeatures != 0) {
                         if (count == 0) {
                             mBuilder.append("M  RBC");
@@ -469,7 +469,7 @@ public class MolfileCreator {
             if (no != 0) {
                 int count = 0;
                 for (int atom=0; atom<mol.getAllAtoms(); atom++) {
-                    long substitution = mol.getAtomQueryFeatures(atom) & (Molecule.cAtomQFMoreNeighbours | Molecule.cAtomQFNoMoreNeighbours);
+                    int substitution = mol.getAtomQueryFeatures(atom) & (Molecule.cAtomQFMoreNeighbours | Molecule.cAtomQFNoMoreNeighbours);
                     if (substitution != 0) {
                         if (count == 0) {
                             mBuilder.append("M  SUB");
