+++ src/main/java/com/actelion/research/share/gui/editor/chem/IDepictor.java
@@ -54,5 +54,5 @@ public interface IDepictor<T>
 
     void paint(IDrawContext<T> ctx);
 
-    void setFragmentNoColor(long color);
+    void setFragmentNoColor(int color);
 }
